test(lifecycle): extract permanent migration regressions
