+++ tests/component/codex/CodexBackendLifecycleTest.cpp
@@ -49,7 +49,7 @@ namespace {
     typed::Turn turn(const std::string& id, typed::TurnStatus status) {
         typed::Turn value;
         value.id = typed::TurnId{id};
-        value.threadId = typed::ThreadId{"thread-phase3"};
+        value.threadId = typed::ThreadId{"codex-lifecycle-thread"};
         value.status = std::move(status);
         value.raw = ai::openai::codex::Json::object();
         return value;
@@ -79,7 +79,7 @@ namespace {
 
 int main() {
     tests::support::TestResult result;
-    const std::filesystem::path path = std::filesystem::temp_directory_path() / "snodec-phase3-codex-lifecycle.jsonl";
+    const std::filesystem::path path = std::filesystem::temp_directory_path() / "snodec-codex-backend-lifecycle.jsonl";
     std::error_code error;
     std::filesystem::remove(path, error);
 
@@ -96,7 +96,7 @@ int main() {
 
         reducer.apply(state, backend::ThreadUpserted{thread("thread-hydrated"), backend::EntityLoad::Summary});
         reducer.apply(state, backend::TurnUpserted{turn("turn-hydrated", typed::TurnStatus::inProgress())});
-        applyTyped(typed::Event{typed::ThreadStarted{thread("thread-phase3"), ai::openai::codex::Json::object()}});
+        applyTyped(typed::Event{typed::ThreadStarted{thread("codex-lifecycle-thread"), ai::openai::codex::Json::object()}});
 
         typed::Turn success = turn("turn-success", typed::TurnStatus::inProgress());
         applyTyped(typed::Event{typed::TurnStarted{success, ai::openai::codex::Json::object()}});
@@ -120,21 +120,21 @@ int main() {
     }
 
     const std::vector<nlohmann::json> captured = records(path);
-    result.expectTrue(countMessage(captured, "thread created: thread=thread-phase3") == 1, "thread creation is emitted exactly once");
+    result.expectTrue(countMessage(captured, "thread created: thread=codex-lifecycle-thread") == 1, "thread creation is emitted exactly once");
     result.expectTrue(countMessage(captured, "thread created: thread=thread-hydrated") == 0,
                       "generic thread hydration does not fabricate creation");
-    result.expectTrue(countMessage(captured, "turn started: thread=thread-phase3 turn=turn-hydrated") == 0,
+    result.expectTrue(countMessage(captured, "turn started: thread=codex-lifecycle-thread turn=turn-hydrated") == 0,
                       "generic turn hydration does not fabricate a start");
-    result.expectTrue(countMessage(captured, "turn started: thread=thread-phase3 turn=turn-success") == 1, "successful turn has one start");
-    result.expectTrue(countMessage(captured, "turn completed: thread=thread-phase3 turn=turn-success") == 1,
+    result.expectTrue(countMessage(captured, "turn started: thread=codex-lifecycle-thread turn=turn-success") == 1, "successful turn has one start");
+    result.expectTrue(countMessage(captured, "turn completed: thread=codex-lifecycle-thread turn=turn-success") == 1,
                       "successful turn has one completion");
-    result.expectTrue(countMessage(captured, "turn failed: thread=thread-phase3 turn=turn-failed") == 1,
+    result.expectTrue(countMessage(captured, "turn failed: thread=codex-lifecycle-thread turn=turn-failed") == 1,
                       "failed turn has one failure and no later completion");
-    result.expectTrue(countMessage(captured, "turn completed: thread=thread-phase3 turn=turn-failed") == 0,
+    result.expectTrue(countMessage(captured, "turn completed: thread=codex-lifecycle-thread turn=turn-failed") == 0,
                       "failed turn cannot later complete");
-    result.expectTrue(countMessage(captured, "turn cancelled: thread=thread-phase3 turn=turn-cancelled") == 1,
+    result.expectTrue(countMessage(captured, "turn cancelled: thread=codex-lifecycle-thread turn=turn-cancelled") == 1,
                       "cancelled turn has one cancellation");
-    result.expectTrue(countMessage(captured, "turn completed: thread=thread-phase3 turn=turn-cancelled") == 0,
+    result.expectTrue(countMessage(captured, "turn completed: thread=codex-lifecycle-thread turn=turn-cancelled") == 0,
                       "cancelled turn cannot later complete");
 
     for (const auto& record : captured) {
